feat: add logic for modules
